Flipped the logic of the database work. Important commit, please read:
What this essentially does is change the filtering so that when items
are entered, they do not keep a domain root. This is so it can be
changed at code runtime instead of being stored in the database, this
offers a little more flexibility for serving a single domain.

So, in summary:

* Get methods were changed to strip the root domain component and then
query, instead of ensuring it existed before querying.
* Other CRUD methods no longer have the root domain information present
or required; they do not need to be present in the database.

Signed-off-by: Erik Hollensbe <github@hollensbe.org>